修改了一些bug 优化购物流程

+++ Lib/Model/CartModel.class.php
@@ -8,9 +8,15 @@
  * @lastupdate 2010-11-25
  */
 class CartModel extends Model {
+
+	public $memberID;
+	function _initialize() {
+		parent::_initialize();
+		$this->memberID=Session::get('memberID');//会员帐号
+	}
 	// 查验物品
 	public function check_item($sessionID, $pid,$model) {
-		$list = $this->where ( "session_id='" . $sessionID . "' and pid=" . $pid." and model='".$model."'" )->find ();
+		$list = $this->where ( "(uid='".$this->memberID."' or session_id='" . $sessionID . "') and pid=" . $pid." and model='".$model."'" )->find ();
 		if (! $list) {
 			return 0;
 		} else {
@@ -20,52 +26,56 @@ class CartModel extends Model {
 	}
 	//添加购物车
 	public function add_item($sessionID, $pid, $count,$model) {
+		$memberID=Session::get('memberID')?Session::get('memberID'):"not a member";
 		$qty = $this->check_item ( $sessionID, $pid,$model );
-		
+
 		if ($qty) {
 			$data ['count'] = array ('exp', 'count+' . $count );
 			$data['dateline']=time();
-			$this->where ( "session_id='" . $sessionID . "' and pid=" . $pid." and model='".$model."'" )->save ( $data );
+			$this->where ( "(uid='".$memberID."' or session_id='" . $sessionID . "') and pid=" . $pid." and model='".$model."'" )->save ( $data );
 		} else {
 			$data['pid']=$pid;
+			$data['uid']=$this->memberID;
 			$data['session_id']=$sessionID;
 			$data['count']=$count;
 			$data['model']=$model;
 			$data['dateline']=time();
+
 			$this->add($data);
 		}
-	
+
 	}
 	// 删除购物车中的内容
 	public function delete_item($sessionID, $id){
-		$this->where ( "session_id='" . $sessionID . "' and id=" . $id)->delete();
+		$this->where ( "(uid='".$this->memberID."' or session_id='" . $sessionID . "') and id=" . $id)->delete();
 	}
 	//修改物品数量
 	public function modify_quantity($sessionID, $id,$count,$model){
 		$data['count']=$count;
-		$data['model']=$model;
+		//$data['model']=$model;
 		$data['dateline']=time();
-		$this->where ( "session_id='" . $sessionID . "' and id=" . $id )->save($data);
+		$this->where ( "(uid='".$this->memberID."' or session_id='" . $sessionID . "') and id=" . $id )->save($data);
 	}
 	//获取产品种类数
 	public function get_item_count($sessionID){
-		return $this->where ( "session_id='" . $sessionID."'"  )->count();
+	
+		return $this->where ( "(uid='".$this->memberID."' or session_id='" . $sessionID . "')"  )->count();
 	}
 	//获取产品总数
 	public function get_item_totalcount($sessionID){
-		return $this->where ( "session_id='" . $sessionID."'"  )->sum('count');
+		return $this->where ( "(uid='".$this->memberID."' or session_id='" . $sessionID . "')"  )->sum('count');
 	}
 	//清空购物车
 	public function clear_cart($sessionID){
-		$this->where ( "session_id='" . $sessionID."'"  )->delete();
+		$this->where ( "(uid='".$this->memberID."' or session_id='" . $sessionID . "')"  )->delete();
 	}
 	//获取总价格
 	public function cart_total($sessionID){
 		$total=0;
 		$dao=D("Products");
-		$list=$this->where( "session_id='" . $sessionID."'")->findAll();
+		$list=$this->where( "(uid='".$this->memberID."' or session_id='" . $sessionID . "')")->findAll();
 		for ($row=0;$row<count($list);$row++){
-			$data=$dao->getpriceInfo($list[$row]['pid'],$list[$row]['count']);
+			$data=$dao->getpriceInfo($list[$row]['pid'],$list[$row]['count'],unserialize($list[$row]['model']));
 			$total=$total+$data['total'];
 		}
 		return $total;
@@ -74,30 +84,32 @@ class CartModel extends Model {
 	public function cart_total_weight($sessionID){
 		$total = 0;
 		$dao=D("Products");
-		$list = $this->where ( "session_id='" . $sessionID . "'" )->findAll ();
+		$list = $this->where ( "(uid='".$this->memberID."' or session_id='" . $sessionID . "')" )->findAll ();
 		for($row = 0; $row < count ( $list ); $row ++) {
 			$weight = $dao->get_weight($list[$row]["pid"])*$list [$row] ['count'] ;
 			$total = $total + $weight;
 		}
 		return $total;
-		
+
 	}
 	//获取购物车明细
 	public function display_contents($sessionID){
-		$list = $this->where (  "session_id='" . $sessionID."'")->findAll ();
+		$list = $this->where (  "session_id='" . $sessionID."' or uid='".$this->memberID."'")->findAll ();
 		if (!$list){
 			return null;
 		}
 		$dao = D ( "Products" );
 		for($row = 0; $row < count ( $list ); $row ++) {
-			$data = $dao->getpriceInfo ( $list [$row] ['pid'], $list [$row] ['count'] );
+
+			$list[$row]['model']= unserialize($list[$row]['model']) ;
+			$data = $dao->getpriceInfo ( $list [$row] ['pid'], $list [$row] ['count'],$list[$row]['model'] );
+
 			$list[$row]['price']=$data['price'];
 			//$list[$row]['pid']=$data['id'];
 			$list[$row]['name']=$data['name'];
 			$list[$row]['bigimage']=$data['bigimage'];
 			$list[$row]['smallimage']=$data['smallimage'];
-			$list[$row]['model']= unserialize($list[$row]['model']) ;
-			$list[$row]['model2']= $dao->get_attrs($dao->where(array('id'=>$list[$row]['pid']))->getField('cateid'),$list[$row]['pid']); ;
+			//$list[$row]['model2']= $dao->get_attrs($dao->where(array('id'=>$list[$row]['pid']))->getField('cateid'),$list[$row]['pid']);
 			//$list[$row]['model']= $list['model'] ;
 			$list[$row]['pricespe']=$data['pricespe'];
 			$list[$row]['count']=$data['count'];
@@ -105,8 +117,10 @@ class CartModel extends Model {
 			$list[$row]['price_total']=$data['price_total'];
 			$list[$row]['pricespe_total']=$data['pricespe_total'];
 			$list[$row]['total']=$data['total'];
+			$list[$row]['costprice']=$data['costprice'];
+			$list[$row]['provider']=$data['provider'];
 		}
-		
+
 		return $list;
 	}
 }
